pyproject.toml
# pyproject.toml
[project]
name = "nanobot-channel-deltachat"
version = "0.1.0"
dependencies = ["nanobot-ai","deltachat-rpc-client"]

[project.entry-points."nanobot.channels"]
deltachat = "nanobot_channel_deltachat:DeltaChatChannel"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot_channel_deltachat"]